duke package: separate the package into two packages
Previous duke package contains too many classes that serve different purposes

Add an additional package called task and place Task class and its child classes into the package.

Improve clarity in purpose of each package.
