+++ src/main/java/duke/Storage.java
@@ -1,5 +1,10 @@
 package duke;
 
+import task.Deadline;
+import task.Event;
+import task.Task;
+import task.ToDos;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
