+++ src/test/java/DukeTest.java
@@ -1,7 +1,8 @@
 //package seedu;
 
-import duke.ToDos;
 import org.junit.jupiter.api.Test;
+import task.ToDos;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class DukeTest {
